com.unity.tiny@0.31.0-preview.23 ## [0.31.0] - 2020-09-24
### Added

* TinyJsonInterface, supports reading / writing json fields.
* TinyJsonStreamingWriter, performant stack-based json writer.
* 2D conversions
* 2D rendering pipeline for DotsRuntime
* Added support for rendering layers in the 2D rendering pipeline
* UpdateWorldTimeSystem to Tiny namespace
* ElapsedTime and DeltaTime properties to UnityInstance
* hook UnityInstance into worlds with UpdateWorldTimeSystem, after the Loading phase
* Initial support for custom shaders.

### Removed

* The `Unity.Tiny.Scenes` asmdef has been removed. Please prefer `Unity.Scenes` instead as it is tiny compatible, and provides the same familiar Unity workflow when exporting scenes (i.e. you no longer require a specific DOTS Subscene to store your content. We now export the main scene as well as any additional subscenes appropriately)
* World.SetTime from UnityInstance Running phase

### Fixed

* Crash in TinyTime sample on Windows
* Crash during shader compilation when using Unity 2020.2
* Skinned meshes don't work with unlit shaders

## [0.30.0] - 2020-08-26

### Added

* Suspend and resume game time in the runtime
* Multiple cameras are converted properly, respecting depth and view rectangle
* CameraMask component is converted for both mesh renderers and cameras. For cameras culling mask is converted, for mesh renderers layer.

### Changed

* TinyAnimation no longer discards empty animation clips. If a clip becomes empty during conversion, a warning will be displayed, but the clip will be preserved.
* Cameras now require the CameraAutoAspectFromNode component to automatically match the render target aspect ratio. The component is automatically added during conversion, but can be removed in case manual aspect ratio control is desired.

### Deprecated

* Unity.Tiny.Utils.StringBuilder and StringFormatter have been removed.  Please use the FixedString formatting

### Removed

* Unity.Tiny.IO and Unity.Tiny.Thread.Native AssemblyDefinitions have been moved from the `com.unity.tiny` package into the `com.unity.dots.runtime` package. Since `com.unity.tiny` relies on `com.unity.dots.runtime` users should not expect any breakages.
* `Unity.Tiny.TinyEnvironment` has been removed. If configuration data is required, users should consider using `GetSingleton<MyConfigComponent>()`/`SetSingleton(myConfigComponent)` instead of `TinyEnvironment.GetConfigData<>()`/TinyEnvironment.SetConfigData()`

### Fixed

* TinyAnimation will now properly apply scale patching when using the `Animation` component without the `TinyAnimationScalePatcher` component.
* TinyAnimation will now properly handle the conversion of animation clips with partial Euler rotation data (having a binding to `localEulerAnglesRaw.y` but not to `.x` or `.z`, for example).
* Fixed the webp decoder fallback to work with managed debugging web builds.
* Text was always being drawn on top of other content.  It will now sort properly against other transparent content (by depth).

## [0.29.0] - 2020-08-04

### Added

* Added APIs to `TinyAnimation` to allow selecting clips by hash or by name. To take advantage of it, please use `TinyAnimation.StringToHash`.

### Deprecated

* `TinyAnimation.SelectClip(World, Entity, int)` was deprecated and renamed `TinyAnimation.SelectClipAtIndex(World, Entity, int)` to more easily differentiate it from the newly introduced `TinyAnimation.SelectClip(World, Entity, uint)`.
* `TinyAnimation.GetDuration(World, Entity, int)` was deprecated and renamed `TinyAnimation.GetDurationAtIndex(World, Entity, int)` to more easily differentiate it from the newly introduced `TinyAnimation.GetDuration(World, Entity, uint)`.

### Fixed

* TinyAnimation no longer causes crashes when instantiating an animated entity from a prefab, if the animation is set to "Play Automatically".
* TinyAnimation now ensures that if a default clip is set on the `Animation` component, it will be the one playing by default.
* Fixed a crash in `TinyAnimation` that happens when animating references and float values on the same entity.
* Fixed an issue in `TinyAnimation` where the `CreateAnimationBinding` attribute would not be able to bind an asset field to an entity field.
* ScreenSpaceToWorldSpace now also works correctly when rendering direct to framebuffer
* ScreenSpaceToWorldSpaceRay is no longer inverted

+++ Unity.Tiny.Audio.Native/cpp~/SoundSource.cpp
@@ -107,19 +107,19 @@ const float* SoundSource::fetchAndResample(uint32_t frameCount, uint32_t* delive
         resampledFramesAvailable = (uint32_t)((numFrames - m_framePosResample) / m_pitch);
 
     uint32_t resampledFrameCount = (frameCount <= resampledFramesAvailable) ? frameCount : resampledFramesAvailable;
-    for (int i = 0; i < resampledFrameCount; i++)
+    for (uint32_t i = 0; i < resampledFrameCount; i++)
     {            
-        uint64_t framePosResample1 = (uint64_t)m_framePosResample;
+        uint64_t framePosResample1 = (uint64_t) m_framePosResample;
         uint64_t framePosResample0 = (framePosResample1 == 0) ? numFrames-1 : framePosResample1-1;
         uint64_t framePosResample2 = framePosResample1+1 >= numFrames ? framePosResample1+1-numFrames : framePosResample1+1;
         uint64_t framePosResample3 = framePosResample1+2 >= numFrames ? framePosResample1+2-numFrames : framePosResample1+2;
 
         for (int iSample = 0; iSample < 2; iSample++)
         {
-            uint32_t samplePosTemp0 = framePosResample0*2 + iSample;
-            uint32_t samplePosTemp1 = framePosResample1*2 + iSample;
-            uint32_t samplePosTemp2 = framePosResample2*2 + iSample;
-            uint32_t samplePosTemp3 = framePosResample3*2 + iSample;
+            uint32_t samplePosTemp0 = (uint32_t) (framePosResample0*2 + iSample);
+            uint32_t samplePosTemp1 = (uint32_t) (framePosResample1*2 + iSample);
+            uint32_t samplePosTemp2 = (uint32_t) (framePosResample2*2 + iSample);
+            uint32_t samplePosTemp3 = (uint32_t) (framePosResample3*2 + iSample);
 
             float mu = (float)(m_framePosResample - (double)framePosResample1);
